fixed: use int16_t for m_pwmphase, should save some casts, fixed: flip all bits instead of doing * -1, will give maximum range
git-svn-id: http://boblight.googlecode.com/svn/trunk@199 c8956a1e-8086-11de-bc80-f7d40095c7a4

+++ src/device/devicesound.h
@@ -49,7 +49,7 @@ class CDeviceSound : public CDevice
     double     m_latency;         //latency in milliseconds
     int        m_period;          //period in samples
  
-    int        m_pwmphase;        //phase of the pwm, gets flipped when pwmcount hits pwmperiod
+    int16_t    m_pwmphase;        //phase of the pwm, gets flipped when pwmcount hits pwmperiod
     int        m_pwmperiod;       //size of the pwm period in samples
     int        m_pwmcount;        //current pwm position
